+++ app/containers/MainContentArea/index.js
@@ -10,6 +10,7 @@
 
 import React from 'react';
 import { connect } from 'react-redux';
+import { injectIntl, intlShape } from 'react-intl';
 import PropTypes from 'prop-types';
 import Radium from 'radium';
 
@@ -23,17 +24,31 @@ import EmailContentArea from 'containers/EmailContentArea';
 import VoiceContentArea from 'containers/VoiceContentArea';
 import WorkItemContentArea from 'containers/WorkItemContentArea';
 import WelcomeStats from 'containers/WelcomeStats';
+import NotificationBanner from 'components/NotificationBanner';
 
-import { setInteractionConfirmation } from 'containers/AgentDesktop/actions';
+import {
+  setInteractionConfirmation,
+  toggleInteractionNotification,
+} from 'containers/AgentDesktop/actions';
 import { getSelectedInteraction } from 'containers/AgentDesktop/selectors';
 
-import { selectMessageTemplates } from './selectors';
+import { addErrorToHistory } from 'containers/Errors/actions';
+
+import {
+  selectMessageTemplates,
+  selectLastInteractionNotification,
+} from './selectors';
+import messages from './messages';
 
 const styles = {
   base: {
     backgroundColor: '#072931',
     color: '#4B4B4B',
     position: 'relative',
+  },
+  notificationBanner: {
+    flex: '1 0 content',
+    alignSelf: 'flex-start',
   },
 };
 
@@ -51,9 +66,28 @@ class MainContentArea extends React.Component {
     }
   };
 
+  dismissErrorInteraction = () => {
+    const {
+      errorDescriptionMessage,
+      ...lastInteractionNotification
+    } = this.props.lastInteractionNotification;
+    const {
+      notifications,
+      ...interactionInfo
+    } = this.props.selectedInteraction;
+    this.props.toggleInteractionNotification(
+      this.props.selectedInteraction.interactionId,
+      lastInteractionNotification
+    );
+
+    this.props.addErrorToHistory({
+      ...lastInteractionNotification,
+      ...interactionInfo,
+    });
+  };
+
   render() {
     const { selectedInteraction } = this.props;
-
     let content = null;
     if (selectedInteraction) {
       if (selectedInteraction.status === 'creating-new-interaction') {
@@ -113,6 +147,22 @@ class MainContentArea extends React.Component {
 
     return (
       <div style={[styles.base, this.props.style]}>
+        {this.props.lastInteractionNotification && (
+          <NotificationBanner
+            key={messages.interactionError.id}
+            id="interaction-error-banner"
+            style={styles.notificationBanner}
+            titleMessage={messages.interactionError}
+            descriptionMessage={
+              this.props.lastInteractionNotification.errorDescriptionMessage
+            }
+            isError={this.props.lastInteractionNotification.isError}
+            dismiss={
+              this.props.lastInteractionNotification.isDismissable &&
+              this.dismissErrorInteraction
+            }
+          />
+        )}
         {content !== null ? content : <WelcomeStats agent={this.props.agent} />}
       </div>
     );
@@ -122,12 +172,16 @@ class MainContentArea extends React.Component {
 const mapStateToProps = (state, props) => ({
   selectedInteraction: getSelectedInteraction(state, props),
   messageTemplates: selectMessageTemplates(state, props),
+  lastInteractionNotification: selectLastInteractionNotification(state, props),
 });
 
 function mapDispatchToProps(dispatch) {
   return {
     setInteractionConfirmation: (interactionId, status) =>
       dispatch(setInteractionConfirmation(interactionId, status)),
+    toggleInteractionNotification: (interactionId, notification) =>
+      dispatch(toggleInteractionNotification(interactionId, notification)),
+    addErrorToHistory: (error) => dispatch(addErrorToHistory(error)),
     dispatch,
   };
 }
@@ -138,11 +192,17 @@ MainContentArea.propTypes = {
   style: PropTypes.array,
   agent: PropTypes.object.isRequired,
   setInteractionConfirmation: PropTypes.func.isRequired,
+  intl: intlShape.isRequired,
+  toggleInteractionNotification: PropTypes.func.isRequired,
+  lastInteractionNotification: PropTypes.object,
+  addErrorToHistory: PropTypes.func.isRequired,
 };
 
 export default ErrorBoundary(
-  connect(
-    mapStateToProps,
-    mapDispatchToProps
-  )(Radium(MainContentArea))
+  injectIntl(
+    connect(
+      mapStateToProps,
+      mapDispatchToProps
+    )(Radium(MainContentArea))
+  )
 );