+++ app/containers/MainContentArea/messages.js
@@ -1,0 +1,8 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  interactionError: {
+    id: 'app.containers.MainContentArea.interactionError',
+    defaultMessage: 'Interaction Error',
+  },
+});