+++ app/containers/MainContentArea/selectors.js
@@ -4,6 +4,9 @@
 
 import { createSelector } from 'reselect';
 
+import { getSelectedInteraction } from 'containers/AgentDesktop/selectors';
+import { generateErrorMessage } from 'utils/errorMessage';
+
 const selectCurrentAgentMessageTemplates = (state) =>
   state.getIn(['agentDesktop', 'userConfig', 'messageTemplates']);
 
@@ -13,4 +16,27 @@ const selectMessageTemplates = createSelector(
     messageTemplates.toJS().filter((template) => template.active === true)
 );
 
-export { selectMessageTemplates };
+// Right now only working for interaction errors
+const selectLastInteractionNotification = createSelector(
+  getSelectedInteraction,
+  (state, props) => props.intl.formatMessage,
+  (selectedInteraction, formatMessage) => {
+    const { notifications } = selectedInteraction;
+    if (notifications) {
+      const lastNotification = notifications.pop();
+      if (lastNotification && lastNotification.isError) {
+        const errorDescriptionMessage = generateErrorMessage(
+          lastNotification,
+          formatMessage
+        );
+        return {
+          ...lastNotification,
+          errorDescriptionMessage,
+        };
+      }
+    }
+    return null;
+  }
+);
+
+export { selectMessageTemplates, selectLastInteractionNotification };